+++ src/reducer.c
@@ -45,7 +45,7 @@
  * information somewhat from its original represenation.
  */
 
-void reducer(int lowfreq, int hifreq ,int axis_length)
+void reducer(gfloat lowfreq, gfloat hifreq ,int axis_length)
 {
 	gint i = 0;
 	gint j = 0;
@@ -59,23 +59,20 @@ void reducer(int lowfreq, int hifreq ,int axis_length)
 	gint highbin = 0;
 	gfloat hertz_per_bin = 0.0;
 	gfloat bins_per_pip = 0.0;
-	/* Ideally this should be rewritten to allow arbritrary frequency 
-	 * spreads in the display. i.e low and high freq should be 
-	 * selectable eventually so that you can zoom into the desired range 
-	 */
 	
+	/* determine the frequency spread per bin, and the low and high bin
+	 * limits for recalculating to fit the number of pips (pixels on screen)
+	 */
 	hertz_per_bin = (float)RATE/(float)nsamp;
 	lowbin = lowfreq/hertz_per_bin;
 	highbin = hifreq/hertz_per_bin;
 
-//	bins_per_pip = ((float)nsamp/(RATE/bandwidth))/(fabs(axis_length));
 	bins_per_pip = ((float)(highbin-lowbin))/fabs(axis_length);
-
-//	printf("hertz/bin %f, lowbin %i, highbin %i bins_per_pip %f\n",hertz_per_bin,lowbin,highbin,bins_per_pip);
 
 	if ((bins_per_pip <= 0.0) || (axis_length <= 0))
 		printf("ERROR!!, bins_per_pip %f, axis_length %i\n",bins_per_pip,axis_length);
 	pip = bins_per_pip;
+	//printf("lowbin %i, highbin %i,  bins_per_pip=%f\n",lowbin,highbin,bins_per_pip);
 
 	j = lowbin;
